added expiry date for pastes

+++ app/Http/Controllers/PastebinController.php
@@ -9,42 +9,78 @@ class PastebinController extends Controller
 {
     public function update(Request $request, $id)
     {
+        # Validate the request data
         $request->validate([
-            'text' => 'required'
+            'text' => 'required',
+            'date' => 'required|date|after:today'
         ]);
 
-        $paste = Paste::where('id', '=', $id)->first();
+        $paste = Paste::find($id);
 
         $paste->text = $request->input('text');
+        $paste->date = $request->input('date');
 
         $paste->save();
 
-        return  view('pastebin.edit-success');
+        return redirect('/pastebin')->with([
+            'alert' => 'Congratulations, your paste is edited successfully.'
+        ]);
     }
 
     public function edit($id)
     {
-        $paste = Paste::where('id', '=', $id)->first();
+        $paste = Paste::find($id);
+
+        if (!$paste) {
+            return redirect('/pastebin')->with([
+                'alert' => 'Paste not found.'
+            ]);
+        }
+
         return view('pastebin.edit')->with('paste', $paste);
     }
 
     public function delete($id)
     {
-        $paste = Paste::where('id', '=', $id)->first();
+        $paste = Paste::find($id);
+
+        if (!$paste) {
+            return redirect('/pastebin')->with([
+                'alert' => 'Paste not found.'
+            ]);
+        }
+
         $paste->delete();
-        return view('pastebin.delete-success');
+
+        return redirect('/pastebin')->with([
+            'alert' => 'Congratulations, your paste is deleted successfully.'
+        ]);
     }
 
-    public function view($id)
+    public function show($id)
     {
-        $paste = Paste::where('id', '=', $id)->first();
-        return view('pastebin.view')->with('paste', $paste);
+        $paste = Paste::find($id);
+
+        if (!$paste) {
+            return redirect('/pastebin')->with([
+                'alert' => 'Paste not found.'
+            ]);
+        }
+
+        return view('pastebin.show')->with('paste', $paste);
     }
 
-    public function show(Request $request)
+    public function index(Request $request)
     {
-        $pastes = Paste::all();
-        return view('pastebin.show')->with('pastes', $pastes);
+        $pastes = Paste::orderBy('date')->get();;
+
+        foreach ($pastes as $paste) {
+            if ($paste->date < date("U")) {
+                $paste->delete();
+            }
+        }
+
+        return view('pastebin.index')->with('pastes', $pastes);
     }
 
     public function create(Request $request)
@@ -52,11 +88,12 @@ class PastebinController extends Controller
         return view('pastebin.create');
     }
 
-    public function save(Request $request)
+    public function store(Request $request)
     {
         # Validate the request data
         $request->validate([
-            'text' => 'required'
+            'text' => 'required',
+            'date' => 'required|date|after:today'
         ]);
 
         # Instantiate a new Book Model object
@@ -65,11 +102,14 @@ class PastebinController extends Controller
         # Set the properties
         # Note how each property corresponds to a field in the table
         $paste->text = $request->input('text');
+        $paste->date = $request->input('date');
 
         # Invoke the Eloquent `save` method to generate a new row in the
         # `books` table, with the above data
         $paste->save();
 
-        return  view('pastebin.create-success');
+        return redirect('/pastebin')->with([
+            'alert' => 'Congratulations, your paste is added successfully.'
+        ]);
     }
 }
